desktop-app/requirements.txt
# Gmail Email Cleanmail - Desktop App Dependencies

# Qt for Python (official Qt binding)
PySide6>=6.5.0

# Google APIs
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0

# Build tools (for creating executable)
pyinstaller>=6.0.0
